refactor: creating BingoBoard component
