--- src/components/TravelBingoGame/TravelBingoGame.css
@@ -1,41 +1,0 @@
-.board {
-  max-width: fit-content;
-}
-
-.board > div:first-child > div:first-child {
-  border-radius: 25px 0 0 0;
-}
-
-.board > div:first-child > div:last-child {
-  border-radius: 0 25px 0 0;
-}
-
-.board > div:last-child > div:first-child {
-  border-radius: 0 0 0 25px;
-}
-
-.board > div:last-child > div:last-child {
-  border-radius: 0 0 25px 0;
-}
-
-.board .header-card,
-.board .card {
-  padding: 8px !important;
-}
-
-.header-card {
-  width: 20%;
-  text-align: center;
-  font-size: 1.5em;
-  font-weight: bold;
-  padding: 10px;
-  background-color: #f0f0f0;
-}
-
-.card {
-  width: 20% !important;
-  cursor: pointer;
-  text-align: center;
-  overflow: hidden;
-  text-overflow: clip;
-}
