+++ src/components/TravelBingoGame/TravelBingoGame.tsx
@@ -1,21 +1,34 @@
 import "./TravelBingoGame.css";
 
 import { Card, Space, Typography } from "antd";
+import { useState } from "react";
+import { useTranslation } from "react-i18next";
 import Markdown from "react-markdown";
 
-import { KeyObject, TravelBingoGameData } from "../../data/interfaces";
+import {
+  Challenge,
+  KeyObject,
+  TravelBingoGameData,
+} from "../../data/interfaces";
+import BingoBoard from "../BingoBoard/BingoBoard";
 import { DynamicIconComponent } from "../DynamicIcon/DynamicIcon";
-import Challenges from "./Challenges";
+import ChallengeModal from "./ChallengeModal";
 import Runs from "./Runs";
 
-const { Paragraph } = Typography;
+const { Paragraph, Title } = Typography;
 
 interface Props {
   game: TravelBingoGameData & KeyObject;
 }
 
 const TravelBingoGame = ({ game }: Props) => {
+  const { t } = useTranslation();
   const { icon, title, shortDescription, challenges, gamePlay } = game;
+
+  const [selectedChallenge, setSelectedChallenge] = useState<Challenge | null>(
+    null,
+  );
+
   return (
     <>
       <Space direction="vertical" size={16}>
@@ -30,12 +43,21 @@ const TravelBingoGame = ({ game }: Props) => {
             </div>
           }
         >
+          <Runs game={game} />
           <Paragraph strong>{shortDescription}</Paragraph>
-          <Challenges challenges={challenges} />
+          <Title level={2} style={{ marginTop: 20 }}>
+            {t("challenges")}
+          </Title>
+          <BingoBoard challenges={challenges} onClick={setSelectedChallenge} />
           <div style={{ marginTop: "8px" }}>
             <Markdown>{gamePlay}</Markdown>
           </div>
-          <Runs game={game} />
+          {selectedChallenge && (
+            <ChallengeModal
+              challenge={selectedChallenge}
+              onClose={() => setSelectedChallenge(null)}
+            />
+          )}
         </Card>
       </Space>
     </>
